# backend/src/main/resources/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Avanzada API",
    "description": "REST API for managing support requests (students, staff, admin).",
    "contact": {
      "name": "Avanzada"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:9000",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/requests": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "List requests",
        "description": "Lists requests applying optional filters. Students only see their own requests.",
        "operationId": "listRequests",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Requests found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Create a new request",
        "description": "Creates a new support request. Students can only create requests for themselves; admins may specify another requester.",
        "operationId": "createRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Request created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "400": {
            "description": "Validation error or invalid data"
          },
          "403": {
            "description": "User not allowed to create request"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "Returns all users. Used mainly to select assignees or requesters.",
        "operationId": "listUsers",
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Users found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/states": {
      "get": {
        "tags": [
          "States"
        ],
        "summary": "List states",
        "description": "Returns all request states ordered by display order.",
        "operationId": "listStates",
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "States found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StateDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/requests/{id}": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Get request by id",
        "description": "Returns a single request by id. Students can only access their own requests.",
        "operationId": "getRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Request found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Request not found"
          },
          "403": {
            "description": "Student trying to access another user's request"
          }
        }
      }
    },
...
    "/api/request-types": {
      "get": {
        "tags": [
          "Request types"
        ],
        "summary": "List request types",
        "description": "Returns all request types that can be used when registering or classifying requests.",
        "operationId": "listRequestTypes",
        "responses": {
          "401": {
            "description": "Missing or invalid JWT"
          },
          "200": {
            "description": "Request types found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestTypeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateRequestDto": {
        "required": [
          "channelId",
          "description",
          "requestTypeId",
          "requestedById"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "requestTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "channelId": {
            "type": "integer",
            "format": "int64"
          },
          "requestedById": {
            "type": "integer",
            "format": "int64"
          },
          "registeredAt": {
            "type": "string"
          }
        }
      },
      "RequestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "registeredAt": {
            "type": "string"
          },
          "requestTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "requestTypeCode": {
            "type": "string"
          },
          "requestTypeName": {
            "type": "string"
          },
          "channelId": {
            "type": "integer",
            "format": "int64"
          },
          "channelCode": {
            "type": "string"
          },
          "channelName": {
            "type": "string"
          },
          "stateId": {
            "type": "integer",
            "format": "int64"
          },
          "stateCode": {
            "type": "string"
          },
          "stateName": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "priorityJustification": {
            "type": "string"
          },
          "requestedById": {
            "type": "integer",
            "format": "int64"
          },
          "requestedByIdentifier": {
            "type": "string"
          },
          "requestedByName": {
            "type": "string"
          },
          "assignedToId": {
            "type": "integer",
            "format": "int64"
          },
          "assignedToIdentifier": {
            "type": "string"
          },
          "assignedToName": {
            "type": "string"
          },
          "closureObservation": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "StateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RequestTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Bearer token obtained via /api/auth/login",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}